% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSimulatorCpp2.R
\name{DatedTree}
\alias{DatedTree}
\title{DatedTree function}
\usage{
DatedTree(
  phylo,
  sampleTimes,
  sampleStates = NULL,
  sampleStatesAnnotations = NULL,
  tol = 1e-06,
  minEdgeLength = 0
)
}
\arguments{
\item{phylo}{A phylogenetic tree of class ape::phylo}

\item{sampleTimes}{A numeric vector providing the times of samples. If named, 
taxon labels will be based on names of the corresponding sample times}

\item{sampleStates}{For models with more than one deme, a matrix must be 
supplied describing the probability of sampling from each deme. Each row 
corresponds to a sample in the same order as sampleTimes. Each column 
corresponds to probability of sampling each deme. Column names should be 
defined which correspond to deme names in the model. Rows should sum to one.}

\item{sampleStatesAnnotations}{Vector of possible discrete character states 
for taxa. If inferring taxon state from label, this provides the possible 
matches for taxon annotations.}

\item{tol}{to do}

\item{minEdgeLength}{to do}
}
\value{
to do
}
\description{
Create a DatedTree class object; it includes heights for each node and other 
  helper variables
}
\examples{
to do
#write example
}
\author{
Erik Volz
}
