% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSimulatorCpp2.R
\name{.phylostructure2newick}
\alias{.phylostructure2newick}
\title{Write a newick string for an ape::phylo structure regardless of internal node order}
\usage{
.phylostructure2newick(o)
}
\description{
Ideally this would be the behavior of ape::write.tree, however that function fails for 
internal node orderings that do not follow an undocumented standard.
}
\examples{
print( system.time( ( tr0 <- rcoal( 10000 ) )))
st0 <- Sys.time() 
tr <- .phylostructure2newick ( tr0 )
st1 <- Sys.time() 
tr1 <- read.tree(text = tr )
print( st1 - st0 )
ape::ltt.plot( tr0 )
ape::ltt.lines( tr1, col = 'red', lty = 2 )
}
