% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colik.pik0.R
\name{colik.pik.fgy}
\alias{colik.pik.fgy}
\title{Compute a structured coalescent likelihood given a dated genealogy and a demographic history in FGY format}
\usage{
colik.pik.fgy(
  tree,
  tfgy,
  timeOfOriginBoundaryCondition = TRUE,
  maxHeight = Inf,
  forgiveAgtY = 1,
  AgtY_penalty = 1,
  returnTree = FALSE,
  step_size_res = 10,
  PL2 = FALSE
)
}
\description{
Compute a structured coalescent likelihood given a dated genealogy and a demographic history in FGY format
}
