\name{phydynR-package}
\alias{phydynR-package}
\alias{phydynR}
\docType{package}
\title{
\packageTitle{phydynR}
}
\description{
\packageDescription{phydynR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{phydynR}
\packageIndices{phydynR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{phydynR}

Maintainer: \packageMaintainer{phydynR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
