% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydynR-package.R
\docType{package}
\name{phydynR-package}
\alias{phydynR}
\alias{phydynR-package}
\title{phydynR: Model-based coalescent simulation and likelihood for phylodynamic inference}
\description{
This package provides functions for defining infectious disease models or ecological process models in terms of ODEs (ordinary differential equations) or SDEs (stochastic differential equations); functions for simulating trees conditional on process models; and functions for computing likelihoods of trees conditional on a process model.
}
\references{
1. Volz, Erik M. Complex population dynamics and the coalescent under 
  neutrality. Genetics 190.1 (2012): 187-201.
  
2. Volz, Erik M., Katia Koelle, and Trevor Bedford. Viral phylodynamics. 
  PLoS Comput Biol 9.3 (2013): e1002947.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://emvolz-phylodynamics.github.io/phydynR/}
}

}
\keyword{internal}
